service/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv  # opcional si quieres variables de entorno
